fix(restructure): restructured the repeated functionality
